requirements.txt
discord.py==2.0.0
aiohttp
bs4
dnspython
jishaku
asyncpg
asyncio